+++ Minibank.Core/Converters/ICurrencyRateProvider.cs
@@ -5,6 +5,6 @@ namespace Minibank.Core.Converters
 {
     public interface ICurrencyRateProvider
     {
-        Task<double> GetCurrencyRateAsync(Currency currency, CancellationToken cancellationToken);
+        Task<decimal> GetCurrencyRateAsync(Currency currency, CancellationToken cancellationToken);
     }
 }
